+++ packages/cli-kit/src/api/graphql/all_orgs.ts
@@ -1,13 +1,26 @@
 import {gql} from 'graphql-request'
 
-export interface AllOrganizationsQuerySchema {
-  organizations: {
+export interface AllOrganizationsQuerySchemaOrganization {
+  id: string
+  businessName: string
+  website: string
+  appsNext: boolean
+  apps: {
     nodes: {
       id: string
-      businessName: string
-      website: string
-      appsNext: boolean
+      title: string
+      apiKey: string
+      apiSecretKeys: {
+        secret: string
+      }[]
+      appType: string
     }[]
+  }
+}
+
+export interface AllOrganizationsQuerySchema {
+  organizations: {
+    nodes: AllOrganizationsQuerySchemaOrganization[]
   }
 }
 
@@ -19,6 +32,17 @@ export const AllOrganizationsQuery = gql`
         businessName
         website
         appsNext
+        apps {
+          nodes {
+            id
+            title
+            apiKey
+            apiSecretKeys {
+              secret
+            }
+            appType
+          }
+        }
       }
     }
   }