+++ packages/cli-kit/src/api/graphql/create_deployment.ts
@@ -1,7 +1,7 @@
 import {gql} from 'graphql-request'
 
 export const CreateDeployment = gql`
-  mutation CreateDeployment($apiKey: String!, $uuid: String!, $bundleUrl: String!, $extensions: [ExtensionSettings!]) {
+  mutation CreateDeployment($apiKey: String!, $uuid: String!, $bundleUrl: String!, $extensions: [ExtensionSettings!]!) {
     deploymentCreate(input: {apiKey: $apiKey, uuid: $uuid, bundleUrl: $bundleUrl, extensions: $extensions}) {
       deployment {
         uuid